So heres the thing: We were doing it all wrong. We now have targets, and triggers - both are lists of target types. If a poison action doesn't require being triggered - such as a gratuitous arp, then when its performed the target data is just immediately added to the targets list (assuming it wasnt already there -- if it was then that node on the list is updated to reflect status).
If a poison action DOES require a trigger - like arp request response or
dhcp discover/request response, its added to the trigger list. these are
still seperate lists but are now closely related.

There was no need for dhcp_options_t to be a list. this has been removed
from the session type. the *next node still needs to be removed from the
actual type in the dhcp.h file but i found it more important to document
thoughts right now than to write the code.

Long story short: triggers are attacks that are based on flags, once a
trigger happens they are acted upon and removed from the trigger list
unless they are targeting everyone (src mac 00..00), or unless they are
DNS. In either of those two conditions they are left on the trigger list
until removed by the caller or at shutdown.

This way when someone calls poison_cycle() - non blocking checks are
made to pcap. these checks determine if a trigger was met, and if so
it acts on it, updates the status of the node on the targets list and
removes the trigger (unless ofcourse dns or targeting everyone).

register_trigger function was added but is waaay incomplete.
register_target function should also be added soon. To make it easier on
the devs both of these should ensure the specified target is on the
targets list, but only the former should add the target to the trigger list.

this refactoring makes everything a lot easier; this cuts down the code
for list management (now there are only two lists - one for each target,
and none for the attack types). this also makes me shit turn purple and
smell like strawberry sherbert. for more details plz view
http://www.youtube.com/watch?v=xJkaKAIl_Fc

Skibideebebop, Chritopher Reeves. Its like almost 9 am and i haven't
slept and for some reason am falling in love with the girl from the 6
underground sneaker pimps video. okie dokie on to my riverdance lessons
and finger painting.